[kubik-19] - Добавляет типы, выносит ui, удаляет modules
