+++ packages/client/src/components/RegisatrationForm/RegistrationForm.scss
@@ -1,44 +1,44 @@
-@import 'src/assets/scss/variables';
+@import '../../assets/scss/variables';
 
-.login-form {
+.registration-form {
   max-width: 500px;
-  height: 587px;
+  height: 807px;
   width: 100%;
   background: $linen;
   border-radius: $border-radius-large;
 
-  &-inner {
+  &__inner {
     padding: 50px;
     height: 100%;
   }
 
-  &-wrapper {
+  &__wrapper {
     display: flex;
     flex-direction: column;
     justify-content: space-between;
     height: 100%;
   }
 
-  &-title {
+  &__title {
     font-size: 26px;
     line-height: 30px;
     text-align: center;
     margin-bottom: 21px;
   }
 
-  &-text {
+  &__text {
     font-size: 20px;
     line-height: 25px;
     font-style: italic;
     text-align: center;
   }
 
-  &-btn {
+  &__btn {
     display: block;
     margin: 47px auto 11px;
   }
 
-  &-info {
+  &__info {
     font-size: 14px;
     line-height: 17px;
     text-align: center;
