+++ packages/client/src/ui/GuiButton/GuiButton.scss
@@ -1,4 +1,4 @@
-@import '../../../assets/scss/variables';
+@import '../../assets/scss/variables';
 
 .gui-button {
   background-color: $sandy-brown;
