+++ packages/client/src/ui/GuiInput/GuiInput.scss
@@ -1,4 +1,4 @@
-@import 'src/assets/scss/variables';
+@import '../../assets/scss/variables';
 
 .gui-input {
   position: relative;
